[LNAC-036] feat(sse): Remove SSE endpoint (#11)
- llm.client.ts
- template.controller.ts
- example.controller.ts
- llm.response-generator.ts
